Move ProcessGenotype class to ofa_xception.py
